# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Decision Engine",
    "description": "Open-source payment gateway routing service by Juspay. Selects the optimal payment processor for each transaction in real-time using success-rate scoring, rule-based routing, and elimination logic.\n\n## Authentication\n\nThis API supports two authentication methods:\n\n1. **JWT Bearer Token** - For dashboard/user sessions\n   - Header: `Authorization: Bearer <jwt_token>`\n   - Obtain via `/auth/login` endpoint\n\n2. **API Key** - For service-to-service/programmatic access\n   - Header: `x-api-key: DE_<64_char_hex>`\n   - Create via `/api-key/create` endpoint\n\nProtected endpoints require one of these authentication methods.",
    "version": "1.4.0",
    "contact": {
      "name": "Juspay",
      "url": "https://github.com/juspay/decision-engine"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://sandbox.hyperswitch.io",
      "description": "Hyperswitch sandbox with x-feature: decision-engine"
    }
  ],
...
  "paths": {
...
    "/update-gateway-score": {
      "post": {
        "operationId": "updateGatewayScore",
        "tags": [
          "Score Feedback"
        ],
        "summary": "Update gateway score",
        "description": "Feed a transaction outcome back into the success-rate model. Call this after every transaction so the engine has accurate SR data for future routing decisions.\n\nA `CHARGED` status increases the gateway's SR; failure statuses (`AUTHENTICATION_FAILED`, `AUTHORIZATION_FAILED`, etc.) decrease it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGatewayScoreRequest"
              },
              "example": {
                "merchantId": "test_merchant",
                "gateway": "stripe",
                "paymentId": "pay_001",
                "status": "CHARGED",
                "gatewayReferenceId": "stripe_ref_001",
                "enforceDynamicRoutingFailure": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Score updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateScoreResponse"
                },
                "example": {
                  "message": "Score updated",
                  "merchant_id": "test_merchant",
                  "gateway": "stripe",
                  "payment_id": "pay_001"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key created via /api-key/create"
      },
      "AdminSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-secret",
        "description": "Admin bootstrap secret for creating merchant accounts"
      }
    },
    "schemas": {
...
      "UpdateGatewayScoreRequest": {
        "type": "object",
        "required": [
          "merchantId",
          "gateway",
          "paymentId",
          "status"
        ],
        "properties": {
          "merchantId": {
            "type": "string",
            "example": "test_merchant"
          },
          "gateway": {
            "type": "string",
            "example": "stripe"
          },
          "paymentId": {
            "type": "string",
            "example": "pay_001"
          },
          "status": {
            "type": "string",
            "enum": [
              "CHARGED",
              "AUTHENTICATION_FAILED",
              "AUTHORIZATION_FAILED",
              "JUSPAY_DECLINED",
              "FAILURE"
            ],
            "example": "CHARGED"
          },
          "gatewayReferenceId": {
            "type": "string",
            "example": "stripe_ref_001"
          },
          "enforceDynamicRoutingFailure": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "UpdateScoreResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "merchant_id": {
            "type": "string"
          },
          "gateway": {
            "type": "string"
          },
          "payment_id": {
            "type": "string"
          }
        }
      },
...
    }
  }
}